+++ lua/gruvsquirrel/colors/boxsquirrel.lua
@@ -2,75 +2,64 @@ local with_meta = require('gruvsquirrel.util.tbl').with_meta
 
 ---@class BoxsquirrelColors
 local M = {
-  scrat_0 = '#1d2021', -- dark
-  scrat_1 = '#282828', -- dark
-  scrat_2 = '#32302f', -- dark
-  scrat_3 = '#3c3836', -- dark
-  scrat_4 = '#1d2021', -- dark
-  scrat_5 = '#1d2021', -- dark
-  scrat_6 = '#3c3836', -- dark
-  scrat_7 = '#32302f', -- dark
-  scrat_8 = '#32302f', -- dark
+  dark0 = '#1d2021',
+  dark1 = '#282828',
+  dark2 = '#32302f',
+  dark3 = '#3c3836',
+  dark4 = '#1d2021',
+  dark5 = '#1d2021',
+  dark6 = '#3c3836',
+  dark7 = '#32302f',
+  dark8 = '#32302f',
 
-  rocky_0 = '#928374', -- gray
-  rocky_1 = '#524b44', -- gray
-  rocky_2 = '#6c6055', -- gray
-  rocky_3 = '#7d7063', -- gray
-  rocky_4 = '#5a544e', -- gray
-  rocky_5 = '#8c8057', -- gray
-  rocky_6 = '#504a45', -- gray
+  gray0 = '#928374',
+  gray1 = '#524b44',
+  gray2 = '#6c6055',
+  gray3 = '#7d7063',
+  gray4 = '#5a544e',
+  gray5 = '#8c8057',
+  gray6 = '#504a45',
 
-  sally_0 = '#fbf1c7', -- light
-  sally_1 = '#ebdbb2', -- light
-  sally_2 = '#d5c4a1', -- light
+  light0 = '#fbf1c7',
+  light1 = '#ebdbb2',
+  light2 = '#d5c4a1',
 
-  sandy_0 = '#fabd2f', -- yellow
+  yellow0 = '#fabd2f',
 
-  andie_0 = '#ffffff', -- yellow_orange placeholder
+  orange0 = '#fe8019',
+  orange1 = '#ff8013',
+  orange2 = '#e1814d',
+  orange3 = '#fd6e27',
+  orange4 = '#3e1a01',
+  orange5 = '#fea37c',
 
-  hammy_0 = '#fe8019', -- orange
-  hammy_1 = '#ff8013', -- orange
-  hammy_2 = '#e1814d', -- orange
-  hammy_3 = '#fd6e27', -- orange
-  hammy_4 = '#3e1a01', -- orange
-  hammy_5 = '#fea37c', -- orange
+  red0 = '#ff6961',
+  red1 = '#f72a00',
+  red2 = '#9e2a21',
+  red3 = '#440500',
 
-  conker_0 = '#ffffff', -- orange_red placeholder
+  purple0 = '#d3869b',
+  purple1 = '#b16286',
+  purple2 = '#b7a7b7',
+  purple3 = '#8f3f71',
+  purple4 = '#5c2613',
 
-  bucky_0 = '#ff6961', -- red
-  bucky_1 = '#f72a00', -- red
-  bucky_2 = '#9e2a21', -- red
-  bucky_3 = '#440500', -- red
+  blue0 = '#8ec07c',
+  blue1 = '#6bc29f',
+  blue2 = '#378e6e',
+  blue3 = '#273821',
+  blue4 = '#83a598',
+  blue5 = '#151601',
 
-  benny_0 = '#ffffff', -- red_purple placeholder
-
-  surly_0 = '#d3869b', -- purple
-  surly_1 = '#b16286', -- purple
-  surly_2 = '#b7a7b7', -- purple
-  surly_3 = '#8f3f71', -- purple
-  surly_4 = '#5c2613', -- purple
-
-  jojo_0 = '#ffffff', -- purple_blue placeholder
-
-  chitter_0 = '#8ec07c', -- blue
-  chitter_1 = '#6bc29f', -- blue
-  chitter_2 = '#378e6e', -- blue
-  chitter_3 = '#273821', -- blue
-
-  skippy_0 = '#83a598', -- blue_green
-  skippy_1 = '#151601', -- blue_green
-
-  doreen_0 = '#b8bb26', -- green
-  doreen_1 = '#282903', -- green
-  doreen_2 = '#a1a420', -- green
-  doreen_3 = '#89aa11', -- green
-  doreen_4 = '#696d32', -- green
-  doreen_5 = '#0c631b', -- green
-  doreen_6 = '#8d7716', -- green
-  doreen_7 = '#63a590', -- green
-  doreen_8 = '#0f664d', -- green
-
-  slappy_0 = '#ffffff', -- green_yellow placeholder
+  green0 = '#b8bb26',
+  green1 = '#282903',
+  green2 = '#a1a420',
+  green3 = '#89aa11',
+  green4 = '#696d32',
+  green5 = '#0c631b',
+  green6 = '#8d7716',
+  green7 = '#63a590',
+  green8 = '#0f664d',
 }
 
 ---@type BoxsquirrelColors