+++ lua/gruvsquirrel/highlight-definitions/gruvsquirrel.lua
@@ -5,8 +5,8 @@ local M = {
   attributes = function()
     local d = require('gruvsquirrel.common.highlight_aliases').get()
     return mapper.highlight_group_mapper({
-      { { g.GruvsquirrelCursorColumnFloat }, d.curcol },
-      { { g.GruvsquirrelInclinePrefix }, d.soup },
+      { { g.GruvsquirrelCursorColumnFloat }, d.bg_dark8 },
+      { { g.GruvsquirrelInclinePrefix }, d.fg_dark0_bold },
     })
   end,
 }