+++ lua/gruvsquirrel/highlight-definitions/lsp.lua
@@ -5,10 +5,10 @@ local M = {
   attributes = function()
     local d = require('gruvsquirrel.common.highlight_aliases').get()
     return mapper.highlight_group_mapper({
-      { { g.LspCodeLens, g.LspCodeLensSeparator }, d.hotdog },
-      { { g.LspSignatureActiveParameter }, d.chipotle },
-      { { g.LspReferenceWrite, g.LspReferenceRead, g.LspReferenceText }, d.cumin },
-      { { g.LspInlayHint }, d.dax },
+      { { g.LspCodeLens, g.LspCodeLensSeparator }, d.fg_light0_bg_dark6_italic },
+      { { g.LspSignatureActiveParameter }, d.fg_green0_bold },
+      { { g.LspReferenceWrite, g.LspReferenceRead, g.LspReferenceText }, d.bg_dark3_bold },
+      { { g.LspInlayHint }, d.fg_gray1_bg_dark5_italic },
     })
   end,
 }